+++ CHANGELOG.md
@@ -5,7 +5,14 @@
 ## [Unreleased]
 
 ## [0.0.1] - 2023-05-13
-- 明源云开发规范代码检查
+- 代码生成器：Entity、EntityDTO、Js代理类
+- 代码规范检查&修复：NewEntity
+
+## [0.0.2] - 2023-05-21
+- 代码规范检查&修复：Controller、AppService、DomainService、Dao、Entity、DataApi、ResourceFilter
+
+## [0.0.3] - 2023-05-28
+- lombok 扩展方法智能提示
 
 [Unreleased]: https://github.com/h598937749/mydevtools/compare/v0.0.1...HEAD